distributed-systems/0 Proyecto DNP/src/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
prometheus-client==0.19.0

# OpenTelemetry Core
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0

# Instrumentaciones
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-pika==0.45b0

# Exportadores opcionales
opentelemetry-exporter-jaeger==1.24.0
opentelemetry-exporter-prometheus==1.12.0rc1